Refactors common functions and updates Starboard
Moves common functions like checkURL, getEmbedColor, setTextInputComponent, and setRowComponent to a dedicated common folder for better organization and reusability.

Updates the Starboard feature to allow editing of existing starboard messages. This avoids creating duplicate messages when the original message is edited. Also fixes an issue where the starboard message could not be found if the message content was an embed.